Handle all possible states of the connected openHAB installation in index.ejs (#33)
Currently, the index page assumes, that the openHAB version is
online, without rechecking this.

With this commit, the openhab-cloud now checks, if there even is a connected
openHAB installation with the current account (app.js says, that it can be
undefined). If there's no installation, an appropriate message with a link
to the account page is printed. If the installation is online, the already
known messages are printed according to the openHAB version. Otherwise (if
the installation is offline) an appropriate error message is printed.

Fixes #28

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>