Do not expose, if an account exists for an e-mail address with /lostpassword
Instead of having a separate message for an existing and non-existing account for
a specific e-mail address, openhab-cloud now shows the same message if there was no
technical error. The user now sees the same message for the lostpassword function,
no matter if an account exists for the given e-mail or not.

Fixes #76

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>