Add a new system module which holds system configuration (#81)
Currently only the host, port and protocol information, which are accessible
through the provided getter functions, but it can provide access to other
configurations and settings, too.

This improves the handling of the host, port and protocol part of a base url,
which was currently handled by a single configuration variable, only. With this
it was unclear, what the configuration holds and what is expected (with protocol,
or without, with port or without, and so on). The configuration also slightly
changed. The system object know provides a separate way for specifying the base
url with the new options:
 - protocol: The protocol of the "public" accessible URL for this service (e.g. http
   or https)
 - host: The host- or domain name from which this service is accesible, e.g. localhost
   or example.com
 - port: The port over which this service is accesible, e.g. 80 or 443.

This also fixes #45 by using the new getBaseURL() method of the System class to retrieve
the base url of the service and appends the path to that, instead of a hard-coded URL.