Add QUnit tests for system and fix backward-compatibility issue (#87)
This commit adds tests with QUnit which can be run with npm test. It
also adds the first testing methods for the system module and fixes one
error found by these tests:
WHen using a baseurl (which is supported for backward-compatibility reasons)
without a protocol, http:// will be added for correct parsing of the url,
however, http: was added only.