Expose configured public Google Cloud Messaging settings
This commit adds a new api endpoint, which will expose the google cloud
messaging sender ID which is needed to enable push notifications with
the app.

The sender ID can differ between openHAB-cloud instances, which makes the
implementation in the app a bit more difficult, therefore this endpoint is
a pre-step to make the app's push notification more dynamic.

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>