Design changes to match openHAB
This is the try to get the design a bit more aligned with
openHAB. The new design parts are oriented on the designs of
openHAB2, however, they're not the same. Mostly the header and footer,
as well as the navbar, has changed in this commit. The rounded
elements are replaced by angular ones.

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>