Fix caching issue for authentication of users
Using cache() for the authenticate function probably speeds up the query,
as cached results will be returned, however, this will result in problems,
when the user cahnged his password. Due to the cache of the query results,
the password is cached, too, and therefore, for a specific amount of time,
or until the app is restarted, the old password will be valid for login,
and not the new one.

Fix this by not caching the query of users for now, until a proper way for
cache invalidation is implemented in mongoose-cache.

Fixes #92

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>