Explain how to register a device and provide links to the apps
Instead of just saying, that no devices are registered in openHAB-cloud,
give an exaplanation how to register a device, too.

This commit also adds links to the apps in the play store and itunes app
store, which ids are configurable to provide a link to a custom build, if
the openHAB-cloud admin has one and/or want to provide it.

Fixes #51