Decouple routes from app.js
This is just a first step, which will result in a cleaner app.js entry
point, which just instructs other, more streamlined, objects when they've
to do the "real" work. Routes are a big part of it.

This commit for now only extracts the routes to a Routes object, which
currently is as big as the app.js routes part and will be minimized in
future commits.