Float the "send message" container to the right
Otherwise, the container will be displayed directly under the aside navigation
container, if it hasn't enough items to "move" the "send message" container
to the right. This is inconsistent behaviour dependant on the number of devices
the user has registered.

With the float to right, the "send message" container is always visible on the
right side of the page.