First steps to decouple routes from app.js (#97)
* Decouple routes from app.js

This is just a first step, which will result in a cleaner app.js entry
point, which just instructs other, more streamlined, objects when they've
to do the "real" work. Routes are a big part of it.

This commit for now only extracts the routes to a Routes object, which
currently is as big as the app.js routes part and will be minimized in
future commits.

* Add new RequestTracker object to track requests to openHAB

* Add new mobile device registration services

These services are used to register a mobile device to the currently logged
in user account. By this, the methods are decoupled from the routes.

* Fix ensureAuthenticated call to ensureRestAuthenticated

* Fix merge error

The function should not be called, but should be passed as a callback
app.all().