Allow one user to have multiple openHAB instances in his settings
With this commit, one user can, if the system administrator of the
openHAB-cloud instance configured it so, have multiple instances of
openHAB in his account, instead of just having one.

When a request needs to be proxied to openHAB, the default behaviour
is, that a random openHAB instance saved in the user settings (it may
or may not be the first one saved in the settings) will be selected to
serve this resource. However, if a specific openHAB instance need or
should be targeted, the user can prefix the requested resource with the
uuid of the openHAB instance. These URLs are also generated by the home-
page of openHAB-cloud, as long as the openHAB server is online.

This is a first step, which still does not work with all provided features
in the openHAB ecosystem:
 * Basic UI and PaperUI are working, other UIs should work, too, except
 * Rest API does not work as expected, it will show the swagger.json result
   of the targeted openHAB server (the ID in the URL), however, testing a
   request fires a request to the randomly selected openHAB server
 * Apps (iOS/Android) will always use the randomly selected openHAB instance
   for now

What are the plans for the future:
 * Providing an API which can be used by clients to get a list of registered
   openHAB installations for the account (e.g. for IFTTT, Android and iOS app,
   ...)
 * Maybe get rid of this stupid regex replacement in the response body

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>